# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/deposit": {
      "post": {
        "tags": [
          "Annotation Data Deposit"
        ],
        "summary": "Deposit single QA pair",
        "description": "Store QA pair to annotation platform with type inference",
        "operationId": "deposit_data_api_v1_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DepositResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deposit/batch": {
      "post": {
        "tags": [
          "Annotation Data Deposit"
        ],
        "summary": "Batch deposit",
        "description": "Deposit multiple QA pairs (1-1000 items)",
        "operationId": "deposit_batch_api_v1_deposit_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DepositBatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_DepositBatchResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepositBatchResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[DepositBatchResponse]"
      },
      "APIResponse_DepositResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepositResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[DepositResponse]"
      },
...
      "DepositBatchRequest": {
        "properties": {
          "data_list": {
            "items": {
              "$ref": "#/components/schemas/DepositRequest"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Data List",
            "description": "Data list"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id",
            "description": "Batch ID (auto-generated if not provided)"
          }
        },
        "type": "object",
        "required": [
          "data_list"
        ],
        "title": "DepositBatchRequest",
        "description": "Batch data deposit request"
      },
      "DepositBatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id",
            "description": "Batch ID"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "success_count": {
            "type": "integer",
            "title": "Success Count",
            "description": "Success count"
          },
          "duplicate_count": {
            "type": "integer",
            "title": "Duplicate Count",
            "description": "Duplicate count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count",
            "description": "Failed count"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DepositResponse"
            },
            "type": "array",
            "title": "Results",
            "description": "Detailed results"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "total",
          "success_count",
          "duplicate_count",
          "failed_count",
          "results"
        ],
        "title": "DepositBatchResponse",
        "description": "Batch data deposit response"
      },
      "DepositRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 2,
            "title": "Question",
            "description": "Question content"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Answer content"
          },
          "source_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Trace Id",
            "description": "Original trace_id"
          },
          "source_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Request Id",
            "description": "Original request_id"
          },
          "source_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Group Id",
            "description": "Session group ID"
          },
          "caller": {
            "type": "string",
            "title": "Caller",
            "description": "Caller",
            "default": ""
          },
          "callee": {
            "type": "string",
            "title": "Callee",
            "description": "Callee",
            "default": ""
          },
          "caller_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Type",
            "description": "Caller type"
          },
          "callee_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callee Type",
            "description": "Callee type"
          },
          "data_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Type",
            "description": "Data type (auto-inferred if not provided)"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "Priority (auto-set if not provided)"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Data category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Data tags"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "description": "Extra data"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer"
        ],
        "title": "DepositRequest",
        "description": "Single data deposit request"
      },
      "DepositResponse": {
        "properties": {
          "data_id": {
            "type": "string",
            "title": "Data Id",
            "description": "Data ID"
          },
          "data_hash": {
            "type": "string",
            "title": "Data Hash",
            "description": "Data hash"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Data status"
          },
          "is_duplicate": {
            "type": "boolean",
            "title": "Is Duplicate",
            "description": "Is duplicate data",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data_id",
          "data_hash",
          "status"
        ],
        "title": "DepositResponse",
        "description": "Data deposit response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}